+++ src/main/java/cz/cvut/fit/niadp/mvcgame/model/gameObjects/LifetimeLimitedGameObject.java
@@ -14,7 +14,7 @@ public abstract class LifetimeLimitedGameObject extends GameObject {
          this.bornAt = LocalDateTime.now();
     }
 
-    public long getAge() {
+    public double getAge() {
         return ChronoUnit.MILLIS.between(bornAt, LocalDateTime.now());
     }
